Moved helper functions to more appripriate locations
