+++ src/components/mgt-person/mgt-person.graph.ts
@@ -1,4 +1,5 @@
-import { Contact } from '@microsoft/microsoft-graph-types';
+import { Contact, Person } from '@microsoft/microsoft-graph-types';
+import { findPerson } from '../../graph/graph.people';
 import { IGraph } from '../../IGraph';
 import { prepScopes } from '../../utils/GraphHelpers';
 
@@ -17,4 +18,18 @@ export async function findContactByEmail(graph: IGraph, email: string): Promise<
     .middlewareOptions(prepScopes(scopes))
     .get();
   return result ? result.value : null;
+}
+
+/**
+ * async promise, returns Graph contact and/or Person associated with the email provided
+ * Uses: Graph.findPerson(email) and Graph.findContactByEmail(email)
+ *
+ * @param {string} email
+ * @returns {(Promise<Array<Person | Contact>>)}
+ * @memberof Graph
+ */
+export function findUserByEmail(graph: IGraph, email: string): Promise<Array<Person | Contact>> {
+  return Promise.all([findPerson(graph, email), findContactByEmail(graph, email)]).then(([people, contacts]) => {
+    return ((people as any[]) || []).concat(contacts || []);
+  });
 }
