+++ src/graph/graph.photos.ts
@@ -1,6 +1,7 @@
 import { ResponseType } from '@microsoft/microsoft-graph-client';
 import { IGraph } from '../IGraph';
-import { blobToBase64, prepScopes } from '../utils/GraphHelpers';
+import { prepScopes } from '../utils/GraphHelpers';
+import { blobToBase64 } from '../utils/Utils';
 
 /**
  * retrieves a photo for the specified resource.
